feat: add configurable iavl cache size (backport #10561) (#10698)
* feat: add configurable iavl cache size (#10561)

<!--
The default pull request template is for types feat, fix, or refactor.
For other templates, add one of the following parameters to the url:
- template=docs.md
- template=other.md
-->

Closes: #1714

Bump default cache size to 50mb from 10kb. Allow node operators to set cache size.

---

*All items are required. Please add a note to the item if the item is not applicable and
please add links to any relevant follow up issues.*

I have...

- [ ] included the correct [type prefix](https://github.com/commitizen/conventional-commit-types/blob/v3.0.0/index.json) in the PR title
- [ ] added `!` to the type prefix if API or client breaking change
- [ ] targeted the correct branch (see [PR Targeting](https://github.com/cosmos/cosmos-sdk/blob/master/CONTRIBUTING.md#pr-targeting))
- [ ] provided a link to the relevant issue or specification
- [ ] followed the guidelines for [building modules](https://github.com/cosmos/cosmos-sdk/blob/master/docs/building-modules)
- [ ] included the necessary unit and integration [tests](https://github.com/cosmos/cosmos-sdk/blob/master/CONTRIBUTING.md#testing)
- [ ] added a changelog entry to `CHANGELOG.md`
- [ ] included comments for [documenting Go code](https://blog.golang.org/godoc)
- [ ] updated the relevant documentation or specification
- [ ] reviewed "Files changed" and left comments if necessary
- [ ] confirmed all CI checks have passed

*All items are required. Please add a note if the item is not applicable and please add
your handle next to the items reviewed if you only reviewed selected items.*

I have...

- [ ] confirmed the correct [type prefix](https://github.com/commitizen/conventional-commit-types/blob/v3.0.0/index.json) in the PR title
- [ ] confirmed `!` in the type prefix if API or client breaking change
- [ ] confirmed all author checklist items have been addressed
- [ ] reviewed state machine logic
- [ ] reviewed API design and naming
- [ ] reviewed documentation is accurate
- [ ] reviewed tests and test coverage
- [ ] manually tested (if applicable)

(cherry picked from commit 76dde985364729222a53df33ffe14a0f60d46b7a)

* fix conflicts

* fix conflicts++

* Update CHANGELOG.md

+++ server/config/config.go
@@ -77,7 +77,7 @@ type BaseConfig struct {
 	InterBlockCacheSize int `mapstructure:"inter-block-cache-size"`
 
 	// IAVL cache size; bytes size unit
-	IAVLCacheSize int `mapstructure:"iavl-cache-size"`
+	IAVLCacheSize uint64 `mapstructure:"iavl-cache-size"`
 
 	// Bech32CacheSize is the maximum bytes size of bech32 cache (Default : 1GB)
 	Bech32CacheSize int `mapstructure:"bech32-cache-size"`
